migrate_db.py
from app import app, db
from sqlalchemy import text

def run_migration():
    print("Running database migration (SQLite setup)...")
    
    with app.app_context():
        # For SQLite with flask_sqlalchemy, db.create_all() typically handles schema creation.
        # The previous PostgreSQL-specific migration logic is commented out.
        # If you have specific SQLite migration needs later, you can add them here.
        
        # check_sql = """
        # SELECT column_name FROM information_schema.columns 
        # WHERE table_name = 'available_time_slot' AND column_name = 'municipality';
        # """
        # result = db.session.execute(text(check_sql)).fetchone()
        
        # if not result:
        #     # Add the new columns
        #     print("Adding new columns to available_time_slot table...")
            
        #     alter_sql = """
        #     ALTER TABLE available_time_slot 
        #     ADD COLUMN municipality VARCHAR(100),
        #     ADD COLUMN station_name VARCHAR(100),
        #     ADD COLUMN postal_code VARCHAR(10);
        #     """
            
        #     db.session.execute(text(alter_sql))
        #     db.session.commit()
            
        #     print("Migration completed successfully!")
        # else:
        #     print("Columns already exist, no migration needed.")
        print("For SQLite, ensure db.create_all() in app.py handles initial table creation.")
        print("Migration script adjusted for SQLite. No specific column alterations performed by default.")

if __name__ == "__main__":
    run_migration()